+++ client/branded/src/search-ui/components/CommitSearchResult.module.scss
@@ -1,0 +1,5 @@
+.commit-oid {
+    background: var(--code-bg);
+    display: inline-block;
+    padding: 0.25rem;
+}
