+++ client/branded/src/search-ui/components/ResultContainer.module.scss
@@ -67,3 +67,63 @@
         background-color: var(--subtle-bg-2);
     }
 }
+
+.copy-button {
+    visibility: hidden;
+    padding-top: 0;
+    padding-bottom: 0;
+    display: inline-flex;
+    align-items: center;
+
+    .copy-button-container:hover &,
+    .copy-button-container:focus-within & {
+        visibility: visible;
+    }
+}
+
+.divider-between {
+    > *:not(:last-child)::after {
+        content: ' ';
+        height: 1rem;
+        margin: 0 0.75rem;
+        border-right: 1px solid var(--border-color-2);
+        display: block;
+    }
+}
+
+.match-type {
+    white-space: nowrap;
+}
+
+.divider-vertical {
+    border-bottom: 1px solid var(--border-color-2);
+    width: 100%;
+    margin: 0.5rem 0;
+}
+
+.divider {
+    border-right: 1px solid var(--border-color-2);
+    height: 1rem;
+    margin: 0 0.5rem;
+}
+
+// Only used by RepoSearchResult. We should just give that its own class
+.gap-1 {
+    gap: 0.5rem;
+}
+
+.title-inner {
+    overflow-wrap: anywhere;
+}
+
+.icon {
+    margin-right: 0.25rem;
+}
+
+.title {
+    display: flex;
+    align-items: center;
+    flex-grow: 1;
+    min-width: 0;
+    flex-wrap: wrap;
+}
