+++ client/branded/src/search-ui/components/index.ts
@@ -1,5 +1,4 @@
 import FileMatchChildrenStyles from './FileMatchChildren.module.scss'
-import SearchResultStyles from './SearchResult.module.scss'
 import SymbolSearchResultStyles from './SymbolSearchResult.module.scss'
 
 export * from './CodeExcerpt'
@@ -14,9 +13,8 @@ export * from './RepoMetadata'
 export * from './RepoSearchResult'
 export * from './FilePathSearchResult'
 export * from './SymbolSearchResult'
-export * from './ResultContainer'
 export * from './SearchResultStar'
 export * from './SyntaxHighlightedSearchQuery'
 export * from './codeLinkNavigation'
 
-export { SearchResultStyles, FileMatchChildrenStyles, SymbolSearchResultStyles }
+export { FileMatchChildrenStyles, SymbolSearchResultStyles }
